feat: GoogleTest 연동 [GTEST-1]

+++ smart_city_socket_server/cctv/crime_prevention/cp_cctv_command_handler.c
@@ -10,13 +10,13 @@ void cp_cctv_dummy (void *data)
 
 void cp_cctv_id_issuance_handler (void *data)
 {
-    printf("아이디(세션) 발급 커맨드: %d\n", *((uint8_t *)data));
+    printf("아이디(세션) 발급 커맨드: %d\n", *((int *)data));
 }
 
 void cp_cctv_disturb_handler (void *data)
 {
-    printf("방범 서브 커맨드: %d\n", *((uint8_t *)data));
-    uint8_t data_pkt = *((uint8_t *)data);
+    printf("방범 서브 커맨드: %d\n", *((int *)data));
+    uint8_t data_pkt = *((int *)data);
 
     if (data_pkt == ON)
     {
